useModal로 모달 열기 닫기 상태관리 기능 추가
